Problem and Problem 2 Solutions.
All the code is commented out. Since all the code will go in one page, I'll comment all of it out. That way it will be easier to go direct to the problem and get the exact code. Or maybe I could have the solutions in groups of 10. Or 5. I'll see.